refactor: Handle store interception in CozyWebView
Instead of in the HomeView
We want everything intercepted in CozyWebView for better readability.

It is still possible to override this behavior by giving
OnShouldStartLoadWithRequest to CozyWebView component
